Split out mapSearchParamCharacteristicsForUi
In order to break dependency cycles
